[WIP]Add py-spdk client for SPDK
Spdk is written in c, and if the project of python needs to call
Spdk, we need to have a client of python build on the Spdk. The
py-spdk is such a client which uses the python language for
writing so that the python project can call SPDK, such as
Openstack Cyborg. The Cyborg can add a SPDK driver which call
py-spdk to manage SPDK accelerator, and further improve storage
performance.

We need to convert SPDK from c to python through the swig tool,
which will generate the.so file in this transformation process,
and then import the required modules into the py-spdk.

Change-Id: I3d2f0086191d74fb5c71b3b0acbd630a5fda12b7
Signed-off-by: hellowaywewe <simple_hlw@163.com>